let painSelection handle size_t instead of int

+++ src/messages/layouts/MessageLayoutContainer.cpp
@@ -523,7 +523,8 @@ void MessageLayoutContainer::paintAnimatedElements(QPainter &painter,
     }
 }
 
-void MessageLayoutContainer::paintSelection(QPainter &painter, int messageIndex,
+void MessageLayoutContainer::paintSelection(QPainter &painter,
+                                            size_t messageIndex,
                                             const Selection &selection,
                                             int yOffset)
 {
